# src/backendApp/resources/openapi.yaml
openapi: 3.0.0
info:
  title: 'My Sudoku App'
  description: 'This is a pet-project.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
    description: 'API over HTTPS'
paths:
  /api/config:
    get:
      tags:
        - get-data
        - config
      operationId: get_get-config
      responses:
        '200':
          description: 'Successful response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponseDto'
  /api/mercure/publish:
    get:
      tags:
        - testing
        - get-data
        - mercure
        - mercure-publish
      operationId: get_mercure-publish
      responses:
        default:
          description: ''
  '/api/games/sudoku/instances/{gameId}/actions':
    post:
      tags:
        - post-data
        - game-instance-actions
        - game-sudoku
        - game-sudoku-instance-actions
      operationId: post_create-game-sudoku-instance-action
      parameters:
        -
          name: gameId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionDto'
      responses:
        '200':
          description: 'Successful response'
  /api/games/sudoku/instances:
    post:
      tags:
        - post-data
        - game-instances
        - game-sudoku
        - game-sudoku-instances
      operationId: post_create-game-sudoku-instance
      responses:
        '200':
          description: 'Successful response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceCreateResponseDto'
  '/api/games/sudoku/instances/{gameId}':
    get:
      tags:
        - get-data
        - game-instances
        - game-sudoku
        - game-sudoku-instances
      operationId: get_get-game-sudoku-instance
      parameters:
        -
          name: gameId
          in: path
          description: 'Unique identifier for the Sudoku game instance'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Successful response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceGetResponseDto'
components:
  schemas:
    ActionDto:
      required:
        - id
        - timeDiff
        - effects
      properties:
        id:
          type: string
        timeDiff:
          type: integer
        effects:
          type: array
          items: {  }
      type: object
    ConfigResponseDto:
      required:
        - mercurePublicUrl
        - allUsers
        - cachedDatetime
      properties:
        mercurePublicUrl:
          type: string
          maxLength: 500
          minLength: 10
        allUsers:
          type: array
          items: {  }
        cachedDatetime:
          type: string
      type: object
    InstanceCreateResponseDto:
      required:
        - id
      properties:
        id:
          type: string
      type: object
    InstanceGetResponseDto:
      required:
        - id
        - cells
        - groups
      properties:
        id:
          type: string
        cells:
          type: array
          items:
            $ref: '#/components/schemas/CellRowCollectionDto'
        groups:
          type: array
          items:
            $ref: '#/components/schemas/CellGroupDto'
      type: object
    CellRowCollectionDto:
      required:
        - iterator
      properties:
        iterator:
          $ref: '#/components/schemas/Traversable'
      type: object
    CellGroupDto:
      required:
        - id
        - type
        - cells
      properties:
        id:
          type: integer
        type:
          type: string
          enum:
            - ROW
            - COL
            - BLC
        cells:
          type: array
          items:
            $ref: '#/components/schemas/CellDto'
      type: object
    Traversable:
      type: object
    CellDto:
      required:
        - coords
        - value
        - protected
      properties:
        coords:
          type: string
        value:
          type: integer
          maximum: 9
          minimum: 0
        notes:
          type: array
          items:
            type: integer
          default: []
        protected:
          type: boolean
      type: object
tags:
  -
    name: get-data
  -
    name: config
  -
    name: testing
  -
    name: mercure
  -
    name: mercure-publish
  -
    name: post-data
  -
    name: game-instance-actions
  -
    name: game-sudoku
  -
    name: game-sudoku-instance-actions
  -
    name: game-instances
  -
    name: game-sudoku-instances